requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-bcrypt
Pillow
pyjwt
pytz
python-dotenv